Refactor caching to a trait. - Removes unnecessary base behavior from majority of drivers.
