+++ module/VuFind/src/VuFind/ILS/Driver/AbstractBase.php
@@ -27,9 +27,6 @@
  */
 namespace VuFind\ILS\Driver;
 
-use VuFind\Cache\KeyGeneratorTrait;
-use Zend\Cache\Storage\StorageInterface;
-
 /**
  * Default ILS driver base class.
  *
@@ -43,40 +40,12 @@ use Zend\Cache\Storage\StorageInterface;
  */
 abstract class AbstractBase implements DriverInterface
 {
-    use KeyGeneratorTrait;
-
-    /**
-     * Cache for storing ILS data temporarily (e.g. patron blocks)
-     *
-     * @var StorageInterface
-     */
-    protected $cache = null;
-
-    /**
-     * Lifetime of cache (in seconds).
-     *
-     * @var int
-     */
-    protected $cacheLifetime = 30;
-
     /**
      * Driver configuration
      *
      * @var array
      */
     protected $config = [];
-
-    /**
-     * Set a cache storage object.
-     *
-     * @param StorageInterface $cache Cache storage interface
-     *
-     * @return void
-     */
-    public function setCacheStorage(StorageInterface $cache = null)
-    {
-        $this->cache = $cache;
-    }
 
     /**
      * Set configuration.
@@ -91,73 +60,5 @@ abstract class AbstractBase implements DriverInterface
     public function setConfig($config)
     {
         $this->config = $config;
-    }
-
-    /**
-     * Helper function for fetching cached data.
-     * Data is cached for up to $this->cacheLifetime seconds so that it would be
-     * faster to process e.g. requests where multiple calls to the backend are made.
-     *
-     * @param string $key Cache entry key
-     *
-     * @return mixed|null Cached entry or null if not cached or expired
-     */
-    protected function getCachedData($key)
-    {
-        // No cache object, no cached results!
-        if (null === $this->cache) {
-            return null;
-        }
-
-        $fullKey = $this->getCacheKey($key);
-        $item = $this->cache->getItem($fullKey);
-        if (null !== $item) {
-            // Return value if still valid:
-            if (time() - $item['time'] < $this->cacheLifetime) {
-                return $item['entry'];
-            }
-            // Clear expired item from cache:
-            $this->cache->removeItem($fullKey);
-        }
-        return null;
-    }
-
-    /**
-     * Helper function for storing cached data.
-     * Data is cached for up to $this->cacheLifetime seconds so that it would be
-     * faster to process e.g. requests where multiple calls to the backend are made.
-     *
-     * @param string $key   Cache entry key
-     * @param mixed  $entry Entry to be cached
-     *
-     * @return void
-     */
-    protected function putCachedData($key, $entry)
-    {
-        // Don't write to cache if we don't have a cache!
-        if (null === $this->cache) {
-            return;
-        }
-        $item = [
-            'time' => time(),
-            'entry' => $entry
-        ];
-        $this->cache->setItem($this->getCacheKey($key), $item);
-    }
-
-    /**
-     * Helper function for removing cached data.
-     *
-     * @param string $key Cache entry key
-     *
-     * @return void
-     */
-    protected function removeCachedData($key)
-    {
-        // Don't write to cache if we don't have a cache!
-        if (null === $this->cache) {
-            return;
-        }
-        $this->cache->removeItem($this->getCacheKey($key));
     }
 }
